Move GPG signing to profile 'ossrh', delete obsolete 'oss-parent' POM
According to this how-to document, the OSS parent POM is obsolete:
http://central.sonatype.org/pages/apache-maven.html#deprecated-oss-parent

For normal users, when building the product it is not necessary to
  - create JavaDoc JAR,
  - create source JARs,
  - sign artifacts with GPG.
That is why I moved those steps into their own Maven profile now. They
are only needed when deploying to the Sonatype OSS repo and to Maven
Central.